// web/src/api/generated/model/storedEvent.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Condux Control Plane API
 * OpenAPI spec version: v1
 */

export interface StoredEvent {
  eventId: string;
  timestamp: string;
  level: string;
  message: string;
  exceptionType: string;
  exceptionValue: string;
  payload: string;
}